changes the name of a State object from the database

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -20,5 +20,6 @@ if __name__ == "__main__":
     session = Session()
     states = session.query(State).filter(State.name.like('%a%')).all()
     for state in states:
-        session.delete(state)
-        session.commit()
+        if 'a' in state.name:
+            session.delete(state)
+    session.commit()
